refactor!: Moved classes to proper package roles
It doesn't make sense for `Position` and `Size` to be in the root package, as they are mainly helper classes.

Similarly, it does not make sense for `Property` to be in the `utils` package, as a `Property` is a core concept in MenuManager. It is the life-blood of how things work and how it should work. It should live alongside the likes of `MenuManager` in importance.
