+++ lib/src/main/java/examples/DemoEditable.java
@@ -8,7 +8,7 @@ import io.github.bfur64.menu.item.display.StaticText;
 import io.github.bfur64.menu.item.input.InputItem;
 import io.github.bfur64.menu.item.input.KeyInputItem;
 import io.github.bfur64.menu.item.input.ToggleItem;
-import io.github.bfur64.menu.utils.Property;
+import io.github.bfur64.menu.Property;
 import io.github.bfur64.terminal.Terminal;
 import io.github.bfur64.terminal.input.KeyStroke;
 import io.github.bfur64.terminal.input.KeyType;
