+++ lib/src/main/java/io/github/bfur64/menu/item/input/InputItem.java
@@ -3,7 +3,7 @@ package io.github.bfur64.menu.item.input;
 import io.github.bfur64.menu.input.InputHandler;
 import io.github.bfur64.menu.item.Item;
 import io.github.bfur64.menu.utils.ErrorEvent;
-import io.github.bfur64.menu.utils.Property;
+import io.github.bfur64.menu.Property;
 import io.github.bfur64.terminal.input.KeyStroke;
 import io.github.bfur64.terminal.input.KeyType;
 import org.jspecify.annotations.NullMarked;
