+++ lib/src/main/java/io/github/bfur64/menu/item/input/ToggleItem.java
@@ -1,7 +1,7 @@
 package io.github.bfur64.menu.item.input;
 
 import io.github.bfur64.menu.item.Item;
-import io.github.bfur64.menu.utils.Property;
+import io.github.bfur64.menu.Property;
 import org.jspecify.annotations.NullMarked;
 
 @NullMarked
