+++ lib/src/main/java/io/github/bfur64/menu/utils/Position.java
@@ -1,4 +1,4 @@
-package io.github.bfur64.menu;
+package io.github.bfur64.menu.utils;
 
 import org.jspecify.annotations.NullMarked;
