+++ lib/src/main/java/io/github/bfur64/menu/utils/Size.java
@@ -1,4 +1,4 @@
-package io.github.bfur64.menu;
+package io.github.bfur64.menu.utils;
 
 public record Size(int x, int y) {
     public static Size of(int x, int y) {
